backend/requirements.txt
django
django-rest-framework
pylint
autopep8
markdown
django-filter
django-cors-headers
